+++ app/uravn.c
@@ -1,10 +1,10 @@
 #include "myfunc.h"
 #include <math.h>
 
-double *uravn(double a, double b, double c) {
+int *uravn(double a, double b, double c) {
 
-    double x[2];
-    double *f = x;
+    int x[2];
+    int *f = x;
     double d = sqrt(b * b - 4 * a * c);
     x[0] = (-b - d) / 2 / a;
     x[1] = (-b + d) / 2 / a;
