+++ tests/uravn.h
@@ -7,7 +7,7 @@ extern "C" {
 
 TEST(uravn1, suite1)
 {
-    double x[2];
+    int x[2];
     x = uravn(1, 3, -4);
     // tests
     ASSERT_EQ(x[0], -4);
